Proof of concept.
Create a quick proof of concept that shows that we _can_ create the
kinds of GIFs I want to create, even if the code could use some cleanup
and organisational changes.

This just parses the subtitles, finds the appropriate subtitles, and
writes them onto the GIF using the font options specified. It also
supports dynamically picking a font size if no font size is specified.